feat(camera): scenario tests for orbit publish + destroy (issue 430)
